+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.1.0
 
 - Added RegistryModule.remove()
+- Added SearchResult:getFirst()
 
 ## 1.0.0